- create side nav component
